Reference to highcharts object and initalization of highcharts in a seperated factory
Its good that you will be given the option to get a reference to the
highcharts object in your parent scope. That can now be done by using
the chart="myChartRef" attribute.

Also created a new factory that handels the initialization of the
highcharts object and adds the reference to the highcharts object in the
parent scope given that it

a) Doesnt previously exists, if so an error is thrown.
b) The attrbitute chart exists in the directive declaration.